perl: *.pm files should not have the executable bit
The Git::Mail::Address file added in bd869f67b9 ("send-email: add
and use a local copy of Mail::Address", 2018-01-05) had the
executable bit set.  But *.pm files should not have the bit set.  It
breaks nothing but is redundant and confusing as none of the other
files have it, and it's never executed as a stand-alone program.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Reviewed-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>