git-submodule.sh:cmd_update: if submodule branch exists, fetch that instead of default
There are projects such as llvm/clang which use several repositories, and they
might be forked for providing support for various features such as adding Redox
awareness to the toolchain. This typically means the superproject will use
another branch than master, occasionally even use an old commit from that
non-master branch.

Combined with the fact that when incorporating such a hierachy of repositories
usually the user is interested in just the exact commit specified in the
submodule info, it follows that a desireable usecase is to be also able to
provide '--depth 1' or at least have a shallow clone to avoid waiting for ages
for the clone operation to finish.

In theory, this should be straightforward since the git protocol allows
fetching an arbitary commit, but, in practice, some servers do not permit
fetch-by-sha1.

Git submodule seems to be very stubborn and cloning master, although the
wrapper script and the gitmodules-helper could work together to clone directly
the branch specified in the .gitmodules file, if specified.

t7406: add test for non-default branch in submodule: If a submodule uses a
non-default branch and the branch info is versioned, on 'git submodule update
--recursive --init' the correct branch should be checked out.

HISTORY:

TODO:
- move the history out of the commit, but using git notes is a hassle
- add support for intial shallow cloning of submodules if the .gitmodules file
  says the submodule branch should be shallow
  - in case the remote does not support fetch-by-sha1. For example, github does
    not permit it.  This is governed by the uploadpack.allowReachableSHA1InWant
    / uploadpack.allowAnySHA1InWant configuration items.
    - possible solutions:
      - git-fetch knows about --shallow-since or even better shallow-exclude
        which could be set to the (depth+1)-th commit (the boundary commit)
        recorded in the shallow information.
      - fetch the branch specified in .gitmodule as shallow and increase the
        shallowness until the hash is found, or until the entire history is
        fetched and it wasn't found
    - for testing use file:// URL to force --no-local

- fetch into a remote tracking branch?

v5:
- squashed code modifications and test case commits into a single commit
- made the test pass w/o considering a server not supporting fetch-by-sha1

v4:
corrections in t7406 (in separate fixup patches):
- use test_commit instead of echo/add/commit as suggested by Stefan Beller
- use 'git -C' instead of cd .. && git
- cleanup chained submodules after test is done
- don't call init after add, is redundant
- super5 commit is done before submodules chaining
- use super_w instead of the existing super
- change branches in submodules after the link is done

v3:
- new patch:add a test case in t7406 for non-default branch in submodule with 2
  chained submodules, emulating the situation encountered in redox os sources
  which includes the redox-os - llvm - libs submodules chain.

v2:
- added the cover letter details in the commit message
- use git submodule--helper instead of computing in shell the refspec for the
  remote branch

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>